rename Subpage and EditingDesignSubpage components
