+++ src/components/SubpageTabs.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Tabs } from "antd";
 import { SubPage } from "../types";
 import { EditingDesign } from "../templates/disease-cell-line";
-import EditingDesignContent from "./EditingDesign";
+import EditingDesignContent from "./EditingDesignSubpage";
 
 const { container } = require("../style/subpage-tabs.module.css");
